+++ src/main/kotlin/org/example/concertbackend/api/queue/request/RegisterQueueRequest.kt
@@ -1,4 +1,4 @@
-package org.example.concertbackend.api.request
+package org.example.concertbackend.api.queue.request
 
 data class RegisterQueueRequest(
     val userId: Long
