+++ src/main/kotlin/org/example/concertbackend/api/queue/response/RegisterQueueResponse.kt
@@ -1,4 +1,4 @@
-package org.example.concertbackend.api.response
+package org.example.concertbackend.api.queue.response
 
 data class RegisterQueueResponse(
     val token: String,
