+++ src/main/kotlin/org/example/concertbackend/api/queue/response/TokenPositionQueueResponse.kt
@@ -1,4 +1,4 @@
-package org.example.concertbackend.api.response
+package org.example.concertbackend.api.queue.response
 
 data class TokenPositionQueueResponse(
     val token: String,
